// front/src/app/sset/models/eligibility-status.enum.ts
export enum EligibilityStatus {
  PENDING = 'pending',
  NOTELIGIBLE = 'notEligible',
  ELIGIBLE = 'eligible',
  NOMATCH = 'noMatch',
  OUTSIDEPERIOD = 'outsidePeriod',
  OVER365 = 'over365',
  NODISRUPTION = 'noDisruption',
  OAL = 'otherAirlines',
  OUTSIDEAPPR = 'outsideAppr',
  DUPLICATEREQUEST = 'duplicateRequest',
  UNABLEDETERMINATION = 'unableDetermination'
}